dlrm-dncv2/jax_based/requirements.txt
absl-py==2.2.2
astroid==3.3.9
astunparse==1.6.3
attrs==25.3.0
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
chex==0.1.89
clu==0.0.12
dill==0.4.0
distlib==0.3.9
dm-tree==0.1.9
docstring-parser==0.16
einops==0.8.1
etils==1.12.2
fiddle==0.3.0
filelock==3.18.0
flatbuffers==25.2.10
flax==0.10.5
fsspec==2025.3.2
gast==0.6.0
google-pasta==0.2.0
googleapis-common-protos==1.70.0
graphviz==0.20.3
grpcio==1.71.0
h5py==3.13.0
humanize==4.12.2
identify==2.6.9
idna==3.10
immutabledict==4.2.1
importlib-resources==6.5.2
iniconfig==2.1.0
isort==6.0.1
jax==0.6.0
jaxlib==0.6.0
jaxtyping==0.3.1
jinja2==3.1.6
kagglehub==0.3.11
keras==3.9.2
keras-hub==0.20.0
libclang==18.1.1
libcst==1.7.0
markdown==3.8
markdown-it-py==3.0.0
markupsafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
ml-collections==1.1.0
ml-dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
nodeenv==1.9.1
numpy==2.1.3
opt-einsum==3.4.0
optax==0.2.4
optree==0.15.0
orbax-checkpoint==0.11.12
packaging==24.2
platformdirs==4.3.7
pluggy==1.5.0
pre-commit==4.2.0
promise==2.3
protobuf==5.29.4
psutil==7.0.0
pyarrow==19.0.1
pygments==2.19.1
pylint==3.3.6
pytest==8.3.5
pytest-env==1.1.5
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
scipy==1.15.2
setuptools==78.1.0
simple-parsing==0.1.7
simplejson==3.20.1
six==1.17.0
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow-cpu==2.19.0
tensorflow-datasets==4.9.8
tensorflow-metadata==1.17.1
tensorflow-text==2.19.0
tensorstore==0.1.73
termcolor==3.0.1
toml==0.10.2
tomlkit==0.13.2
toolz==1.0.0
torch==2.6.0
tqdm==4.67.1
treescope==0.1.9
typing-extensions==4.13.2
urllib3==2.4.0
virtualenv==20.30.0
wadler-lindig==0.1.5
werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
zipp==3.21.0
